renaming utils package
